{"schema":"decodex/commit/1","summary":"Split status projection modules","authority":"manual"}
